+++ pytorch_lightning/profiler/simple.py
@@ -37,7 +37,6 @@ class SimpleProfiler(BaseProfiler):
         dirpath: Optional[Union[str, Path]] = None,
         filename: Optional[str] = None,
         extended: bool = True,
-        output_filename: Optional[str] = None,
     ) -> None:
         """
         Args:
@@ -53,7 +52,7 @@ class SimpleProfiler(BaseProfiler):
                 If you attempt to start an action which has already started, or
                 if you attempt to stop recording an action which was never started.
         """
-        super().__init__(dirpath=dirpath, filename=filename, output_filename=output_filename)
+        super().__init__(dirpath=dirpath, filename=filename)
         self.current_actions: Dict[str, float] = {}
         self.recorded_durations = defaultdict(list)
         self.extended = extended